Remove query options on recursive call
The `@iot.nextLink` from the server contains the query options from
the original XHR. Setting that as the URL *and* setting the `data`
object in the object passed to JQuery will cause the query options
to be merged, causing the query options to grow as the recursion is
called.

This fixes it by clearing the query options after the first call
and relying on the server to have formed the nextLink to correctly
include the query options.